Description
A few people took part in the computer science competition.

Determine and output the average points of the participants of the Olympiad in the 9th grade, in the 10th grade, in the 11th grade.

Input data

Information about the results of the Olympiad is recorded in a file, each line of which has the form:

last name first name grade

Last name and first name - text strings that do not contain spaces. Class is one of the three numbers 9, 10, 11. The score is an integer from 0 to 100.

In this task, the file must be read line by line, without storing the entire contents of the file in memory.

Output

Output three numbers: average points for class 9, class 10, class 11. The input file is encoded utf-8 (use open ('input.txt', 'r', encoding = 'utf-8')).

Examples

Test 1
Input data:
Ivanov Sergey 9 90
Sergeev Peter 10 91
Petrov Vasily 11 92
Vasilyev Ivan 9 93

The output of the program:
91.5 91.0 92.0