Description
Maximum grade score
A few people took part in the computer science competition. The winner of the competition is the person who scored the most points. Winners are determined independently for each class. Determine the number of points scored by the winner in each class. It is guaranteed that in each class there was at least one participant.

Input format

Information about the results of the Olympiad is recorded in a file, each line of which has the form: last name first name class point.

Last name and first name - text strings that do not contain spaces. Class is one of the three numbers 9, 10, 11. The score is an integer from 0 to 100.

In this task, the file must be read line by line, without storing the contents of the file in memory.

Output format

Output three numbers: the points of the winner of the Olympiad in grade 9, in grade 10, in grade 11. The input file is encoded utf-8 (In Python, use open ('input.txt', 'r', encoding = 'utf-8')).

Test 1
Input:
Иванов Сергей 9 90
Сергеев Петр 10 91
Петров Василий 11 92
Васильев Иван 9 93

Output:
93 91 92