Description
Have you met a number before
The input line contains a sequence of numbers separated by spaces. For each number, print the word YES (on a separate line) if this number was previously encountered in the sequence or NO, if it was not.

Input format

A list of numbers is entered. All numbers in the list are on one line.

Output format

Print the answer to the problem.

Test 1
Input data:
1 2 3 2 3 4

The output of the program:
NO
NO
NO
YES
YES
NO