Description
In the input file (you can read data from sys.stdin by connecting the sys library) the text is written. A word is a sequence of non-blank characters in a row, words separated by one or more spaces or line breaks. Determine how many different words are contained in this text.

Input format

Text is entered.

Output format

Print the answer to the problem.

Examples
Test 1
Input data:
She sells sea shells on the sea shore;
Shells that she sells shells I'm sure.
So if she sells sea shells on the sea shore,
I'm sure that the shells are the sea shore shells.


The output of the program:
19