Description
Each of the N schoolchildren of a certain school knows Mᵢ languages. Determine which languages ​​are known to all students and languages ​​that at least one of the students knows.

Input format

The first line of input contains the number of students N. Next comes the N numbers Mᵢ, after each of the numbers there are Mᵢ lines containing the names of the languages ​​that the i-th student knows. The length of the names of languages ​​does not exceed 1000 characters, the number of different languages ​​is not more than 1000. 1≤N≤1000, 1≤Mᵢ≤500.

Output format

In the first line print the number of languages ​​that all students know. Starting from the second line - a list of such languages. Then - the number of languages ​​that at least one schoolchild knows, on the following lines - a list of such languages.

Examples
Test 1
Input data:
3
3
Russian
English
Japanese
2
Russian
English
one
English

The output of the program:
one
English
3
Russian
Japanese
English