Description
In the input file (you can read data from the file input.txt) the text is written. A word is considered to be a sequence of non-space characters in succession. Words are separated by one or more spaces or line breaks. For each word in this text, count how many times it has been in this text before.

Input format

Text is entered.

Output format

Print the answer to the problem.

Examples

Test 1
Input data:
one two one tho three


The output of the program:
0 0 1 0 0



Test 2
Input data:
She sells sea shells on the sea shore;
Shells that she sells shells I'm sure.
So if she sells sea shells on the sea shore,
I'm sure that the shells are the sea shore shells.


The output of the program:
0 0 0 0 0 0 1 0 0 1 1 0 3 0 2 2 0 0 0 0 1 2 3 3 1 1 0 0 1 1 1 2 2 1 5 0 0



Test 3
Input data:
aba aba; aba @? "

The output of the program:
0 0 1 0