Description
Olympiad results
N people participated in the competition. Each received a certain number of points, while it turned out that all participants had different numbers of points. Sort the list of participants in the competition in descending order of points.

Input format

The program receives as input the number of participants in the Olympiad N. Next comes N lines, each line contains the name of the participant, then, after the space, the number of points he scored.

Output format

Display the list of participants (only last names) in descending order of points.