Description
In the election of the President of the Russian Federation, the candidate who won over half the number of votes wins. If there is no such candidate, then two candidates with the highest number of votes go to the second round of elections.

Input format

Each line of the input file contains the name of the candidate for which one voter voted. It is known that the total number of candidates does not exceed 20, but unlike the previous tasks, the list of candidates is not explicitly specified. Read the data from the input.txt file with the utf8 encoding.

Output format

If there is a candidate who receives more than 50% of the vote, the program should display his name. If there is no such candidate, the program should print the name of the candidate who took the first place, then the name of the candidate who took the second place. Output the data to the file output.txt with indication of utf8 encoding.

Test 1
Input data:
Poluuekt Varfolomeev
Bartholomew Poluektov
Poluuekt Varfolomeev

The output of the program:
Poluuekt Varfolomeev



Test 2
Input data:
Poluuekt Varfolomeev
Bartholomew Vissarionov
Vissarion Poluektov
Bartholomew Vissarionov
Bartholomew Vissarionov
Poluuekt Varfolomeev

The output of the program:
Bartholomew Vissarionov
Poluuekt Varfolomeev